Silences the warnings about sass in sass-loader as I can't do anything about that right now
