+++ libs/components/feedback/loader/index.tsx
@@ -4,7 +4,12 @@ import React from 'react'
 import styles from './styles.module.scss'
 
 interface Props {
-  className?: string
+  /**
+   * Text to display below the loader
+   * only visible to screen readers
+   */
+  helpText?: string
+  className?: string,
 }
 
 /**
@@ -12,15 +17,18 @@ interface Props {
  * @param param0 
  * @returns 
  */
-export const Loader: React.FC<Props> = ({ className = '' }) => {
+export const Loader: React.FC<Props> = ({ className = '', helpText='Loading...' }) => {
   return (
-    <div className={`${styles.loader} ${className}`}>
-      <div className={`${styles.square} ${styles.square1}`}></div>
-      <div className={`${styles.square} ${styles.square2}`}></div>
-      <div className={`${styles.square} ${styles.square3}`}></div>
-      <div className={`${styles.square} ${styles.square4}`}></div>
-      <div className={`${styles.square} ${styles.square5}`}></div>
-    </div>
+    <>
+      <div className={`${styles.loader} ${className}`}>
+        <div className={`${styles.square} ${styles.square1}`}></div>
+        <div className={`${styles.square} ${styles.square2}`}></div>
+        <div className={`${styles.square} ${styles.square3}`}></div>
+        <div className={`${styles.square} ${styles.square4}`}></div>
+        <div className={`${styles.square} ${styles.square5}`}></div>
+      </div>
+      <span className="visually-hidden">{helpText}</span>
+    </>
   )
 }
