+++ libs/components/feedback/loader/styles.module.scss
@@ -50,6 +50,7 @@ $pink-300: hsl(350, 100%, 88%);
   animation-delay: 0.4s;
 }
 
+
 // Bounce animations
 @keyframes bounce1 {
   0%, 100% { transform: translateY(0); }
